Rename utils.kernel_launcher to utils.call_kernel_builder
