+++ numba_dpex/core/kernel_launcher.py
@@ -26,7 +26,7 @@ from numba_dpex.core.types.kernel_api.index_space_ids import (
     NdItemType,
 )
 from numba_dpex.core.types.kernel_api.local_accessor import LocalAccessorType
-from numba_dpex.core.utils import kernel_launcher as kl
+from numba_dpex.core.utils import call_kernel_builder as kl
 from numba_dpex.dpctl_iface import libsyclinterface_bindings as sycl
 from numba_dpex.dpctl_iface.wrappers import wrap_event_reference
 from numba_dpex.kernel_api_impl.spirv.dispatcher import (
